post-processing: Improve YUV to RGB conversion
There are two changes. Firstly we copy the buffer for the asynchronous
thread without converting to RGB as this will hold it up less. But
crucially, it also means that yuvToRgb will get a *cached* rather than
an *uncached* buffer, and this is massively quicker to convert (so the
"unnecessary" copy more than pays for itself).

Secondly, we refactor the yuvToRgb implementation to work on 4x2 pixel
blocks which seems to give a bit more scope for the compiler to
optimise things.

With these changes, converting a VGA buffer goes from about 70ms to
under 6ms (counting both the extra copy and the conversion).

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>